requirements.txt
feedparser
pandas
python-dotenv
os
requests
datetime
email.mime.text
email.mime.multipart
smtplib
